[tests] merge two cases of 9_2_02 into one
