+++ tests/scripts/thread-cert/Makefile.am
@@ -108,8 +108,7 @@ EXTRA_DIST                                                         = \
     Cert_8_1_02_Commissioning.py                                     \
     Cert_8_2_01_JoinerRouter.py                                      \
     Cert_8_2_02_JoinerRouter.py                                      \
-	Cert_9_2_02A_MGMTCommissionerSet.py                              \
-	Cert_9_2_02B_MGMTCommissionerSet.py                              \
+	Cert_9_2_02_MGMTCommissionerSet.py                               \
     Cert_9_2_04_ActiveDataset.py                                     \
     Cert_9_2_07_DelayTimer.py                                        \
     Cert_9_2_08_PersistentDatasets.py                                \
@@ -250,8 +249,7 @@ check_SCRIPTS                                                      = \
     Cert_8_1_02_Commissioning.py                                     \
     Cert_8_2_01_JoinerRouter.py                                      \
     Cert_8_2_02_JoinerRouter.py                                      \
-	Cert_9_2_02A_MGMTCommissionerSet.py                              \
-	Cert_9_2_02B_MGMTCommissionerSet.py                              \
+	Cert_9_2_02_MGMTCommissionerSet.py                               \
     Cert_9_2_04_ActiveDataset.py                                     \
     Cert_9_2_07_DelayTimer.py                                        \
     Cert_9_2_08_PersistentDatasets.py                                \
@@ -293,8 +291,7 @@ XFAIL_NCP_TESTS                                                    = \
     Cert_8_1_02_Commissioning.py                                     \
     Cert_8_2_01_JoinerRouter.py                                      \
     Cert_8_2_02_JoinerRouter.py                                      \
-	Cert_9_2_02A_MGMTCommissionerSet.py                              \
-	Cert_9_2_02B_MGMTCommissionerSet.py                              \
+	Cert_9_2_02_MGMTCommissionerSet.py                               \
     Cert_9_2_04_ActiveDataset.py                                     \
     Cert_9_2_07_DelayTimer.py                                        \
     Cert_9_2_08_PersistentDatasets.py                                \
